+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.14] In progress
 - Added manufacturer field in IMIDIInput and IMIDIOutput interfaces.
 - Ability to listen to only specific devices (based on name / manufacturer).
+- Added pitch bend methods for both input and output.
 
 ## [0.0.13]
 First public (alpha) version.